Issue #64: Only display the current pin on the picture page
Previous attempt to fix #64 introduced another bug.
The map on the picture page now contains info from all images of the album.
On this page we expect to see info (pin/thumbnail) from the current
image only.

This quick fix adds a sql limit condition to the query if the image_id param
exists (I assume that it means we come from the picture page?).
Note that it would be more efficient not to join in this case.