Add the ability to auto center/zoom the map
The "autocenter" behavior was only active for the map displayed on the
category and main pages.
This patch adds an option "Auto center" in the left menu configuration
to use this behavior as well where it could be useful.

Note that this patch also set the default value '0,0' for the center
parameter. Prior to this patch, if it was not filled (empty string),
there was no default value, which caused some errors to be displayed in
the html ("Notice: Undefined variable: center_arr in ...")