Fix #80: Use the default map style for the map v3
If the map v3 is used, the default map style defined in the admin page
is not taken into account.
Actually the issue is deeper: some choices available on the admin page
are not listed in the provider list (OSM FR, OSM HOT, MapQuest for
instance).
The custom map style are not available neither.

This patch adds the following map providers:
- OpenStreetMap_France
- OpenStreetMap_HOT
- MapQuestOpen_OSM
- MapQuestOpen_Aerial
And the custom one.

The one defined as default in the admin form will be activated.
The icons used are retrieved from the providers' servers (tile 5/15/11).