Fix for #87: gpx files should be .gpx$
The osm_get_gps function generates a sql query which retrieves all files
containing "gpx".
Instead it should retrieves files with a filename ending with ".gpx".

Otherwise the map won't display anything because of a JS error
"TypeError: omnivore.png is not a function" (in the case of a png file).