Reset default greenred iconAnchor xy in osm_get_js
Those icons were misplaced on picture map: iconAnchor needed to be updated.
To solve this I copied default values from L.Icon.Default options in leaflet-src.js (3395) https://github.com/xbgmsharp/piwigo-openstreetmap/blob/master/leaflet/leaflet-src.js#L3395
I did the same for popupAnchor because blue default icon has the same footprint as green and red default icons, and because leaflet-src.js is reliable. It's better with 1 x value that let the popup centered, but if -34 y value seems to high for wide maps not that tall, it would be great to move the popup on the side.